ecal added for channel_disabler.py
